chore: update release script

+++ scripts/release.ts
@@ -9,10 +9,10 @@ async function main() {
     type: "select",
     name: "release",
     message: "Select release type",
-    choices: ["patch", "minor", "major", "custom"]
+    choices: ["patch", "minor", "major", "pre", "custom"]
   });
   const pkg = JSON.parse(fs.readFileSync("package.json", "utf-8"));
-  const [major, minor, patch] = pkg.version.split(".");
+  const [major, minor, patch, pre] = pkg.version.split(".");
   switch (release) {
     case "patch":
       pkg.version = [major, minor, Number(patch) + 1].join(".");
@@ -22,6 +22,9 @@ async function main() {
       break;
     case "major":
       pkg.version = [Number(major) + 1, 0, 0].join(".");
+      break;
+    case "pre":
+      pkg.version = [major, minor, patch, Number(pre) + 1].join(".");
       break;
     default:
       {
@@ -35,6 +38,14 @@ async function main() {
       break;
   }
 
+  const confirm = await prompt({
+    type: "confirm",
+    name: "confirm",
+    initial: true,
+    message: `Release v${pkg.version}, Confirm?`
+  });
+  if (!confirm) return;
+
   console.log("Releasing v" + pkg.version + "...");
   fs.writeFileSync("package.json", JSON.stringify(pkg, null, 2) + "\n");
 
@@ -46,8 +57,6 @@ async function main() {
   execSync("git add CHANGELOG.md");
   execSync(`git commit -m "chore: release v${pkg.version}"`);
   execSync(`git tag v${pkg.version}`);
-  execSync(`git push`);
-  execSync(`git push origin --tags v${pkg.version}`);
 }
 
 main();
